Cambios en la practica: Uso de modulos en la calculadora
